dis31: compiles under Linux with GCC
Cosmetic changes to port the source.
Made sure it still compiles with Borland C 3.1 for DOS (for fun)

+++ src/dis31.h
@@ -10,7 +10,7 @@
 
 typedef unsigned char BYTE;
 typedef signed char   SBYTE;
-typedef unsigned int  WORD;
+typedef unsigned short WORD;
 typedef int           BOOL;
 typedef WORD          ADDRESS;
